chore: fix all tests
