+++ games/src/ataxx/tests/mod.rs
@@ -1,4 +1,5 @@
-use ataxx::{perft, Position};
+use crate::ataxx::Position;
+use crate::perft;
 use std::str::FromStr;
 
 macro_rules! perft_test {
@@ -6,7 +7,7 @@ macro_rules! perft_test {
         #[test]
         fn $name() {
             let position = Position::from_str($pos).unwrap();
-            assert_eq!(perft::<true, true>(position, $depth), $nodes)
+            assert_eq!(perft::<true, true, _>(position, $depth), $nodes)
         }
     };
 }
